\name{cubinf.missing}
\alias{anova}
\alias{add1}
\alias{drop1}
\alias{step}

\title{
Missing methods for an object of class "cubinf"
}
\description{
These functions are not implemented for an object of class "cubinf". 
}
\usage{

anova(object, ...)

add1(object, ...)
 
drop1(object, ...)
 
step(object, ...)
 
}

\arguments{
  \item{object}{
An object inheriting from class "glm" or "cubinf".}
  \item{\dots}{
Optional arguments according to the method.}
}

\seealso{
The model fitting function \code{\link{glm}}, \code{\link{cubinf}}
}

\keyword{robust }

